+++ lld/ELF/Relocations.h
@@ -126,6 +126,7 @@ enum RelExpr {
   RE_LOONGARCH_PC_INDIRECT,
   RE_LOONGARCH_TLSGD_PAGE_PC,
   RE_LOONGARCH_TLSDESC_PAGE_PC,
+  RE_HEXAGON_GPREL,
 };
 
 // Architecture-neutral representation of relocation.